Move application middlewares to routers folder
