+++ lib/cartodb/routers/analyses.js
@@ -1,11 +1,11 @@
 const PSQL = require('cartodb-psql');
-const cleanUpQueryParams = require('../middleware/clean-up-query-params');
-const credentials = require('../middleware/credentials');
-const authorize = require('../middleware/authorize');
-const dbConnSetup = require('../middleware/db-conn-setup');
-const rateLimit = require('../middleware/rate-limit');
+const cleanUpQueryParams = require('./middlewares/clean-up-query-params');
+const credentials = require('./middlewares/credentials');
+const authorize = require('./middlewares/authorize');
+const dbConnSetup = require('./middlewares/db-conn-setup');
+const rateLimit = require('./middlewares/rate-limit');
 const { RATE_LIMIT_ENDPOINTS_GROUPS } = rateLimit;
-const cacheControlHeader = require('../middleware/cache-control-header');
+const cacheControlHeader = require('./middlewares/cache-control-header');
 const dbParamsFromResLocals = require('../utils/database-params');
 
 function AnalysesController(pgConnection, authApi, userLimitsApi) {
