+++ lib/cartodb/routers/api-router.js
@@ -40,16 +40,16 @@ const ResourceLocator = require('../models/resource-locator');
 const LayergroupMetadata = require('../utils/layergroup-metadata');
 const RendererStatsReporter = require('../stats/reporter/renderer');
 
-const logger = require('../middleware/logger');
+const logger = require('./middlewares/logger');
 const bodyParser = require('body-parser');
-const servedByHostHeader = require('../middleware/served-by-host-header');
-const stats = require('../middleware/stats');
-const lzmaMiddleware = require('../middleware/lzma');
-const cors = require('../middleware/cors');
-const user = require('../middleware/user');
-const sendResponse = require('../middleware/send-response');
-const syntaxError = require('../middleware/syntax-error');
-const errorMiddleware = require('../middleware/error-middleware');
+const servedByHostHeader = require('./middlewares/served-by-host-header');
+const stats = require('./middlewares/stats');
+const lzmaMiddleware = require('./middlewares/lzma');
+const cors = require('./middlewares/cors');
+const user = require('./middlewares/user');
+const sendResponse = require('./middlewares/send-response');
+const syntaxError = require('./middlewares/syntax-error');
+const errorMiddleware = require('./middlewares/error-middleware');
 
 const MapRouter = require('./map-router');
 const TemplateRouter = require('./template-router');
