+++ lib/cartodb/routers/layergroup/analysis.js
@@ -1,9 +1,9 @@
-const layergroupToken = require('../../middleware/layergroup-token');
-const cleanUpQueryParams = require('../../middleware/clean-up-query-params');
-const credentials = require('../../middleware/credentials');
-const dbConnSetup = require('../../middleware/db-conn-setup');
-const authorize = require('../../middleware/authorize');
-const rateLimit = require('../../middleware/rate-limit');
+const layergroupToken = require('../middlewares/layergroup-token');
+const cleanUpQueryParams = require('../middlewares/clean-up-query-params');
+const credentials = require('../middlewares/credentials');
+const dbConnSetup = require('../middlewares/db-conn-setup');
+const authorize = require('../middlewares/authorize');
+const rateLimit = require('../middlewares/rate-limit');
 const { RATE_LIMIT_ENDPOINTS_GROUPS } = rateLimit;
 const dbParamsFromResLocals = require('../../utils/database-params');
