+++ lib/cartodb/routers/layergroup/dataview.js
@@ -1,15 +1,15 @@
-const layergroupToken = require('../../middleware/layergroup-token');
-const cleanUpQueryParams = require('../../middleware/clean-up-query-params');
-const credentials = require('../../middleware/credentials');
-const dbConnSetup = require('../../middleware/db-conn-setup');
-const authorize = require('../../middleware/authorize');
-const rateLimit = require('../../middleware/rate-limit');
+const layergroupToken = require('../middlewares/layergroup-token');
+const cleanUpQueryParams = require('../middlewares/clean-up-query-params');
+const credentials = require('../middlewares/credentials');
+const dbConnSetup = require('../middlewares/db-conn-setup');
+const authorize = require('../middlewares/authorize');
+const rateLimit = require('../middlewares/rate-limit');
 const { RATE_LIMIT_ENDPOINTS_GROUPS } = rateLimit;
 const createMapStoreMapConfigProvider = require('./middlewares/map-store-map-config-provider');
-const cacheControlHeader = require('../../middleware/cache-control-header');
-const cacheChannelHeader = require('../../middleware/cache-channel-header');
-const surrogateKeyHeader = require('../../middleware/surrogate-key-header');
-const lastModifiedHeader = require('../../middleware/last-modified-header');
+const cacheControlHeader = require('../middlewares/cache-control-header');
+const cacheChannelHeader = require('../middlewares/cache-channel-header');
+const surrogateKeyHeader = require('../middlewares/surrogate-key-header');
+const lastModifiedHeader = require('../middlewares/last-modified-header');
 
 const ALLOWED_DATAVIEW_QUERY_PARAMS = [
     'filters', // json
