+++ lib/cartodb/routers/map/anonymous.js
@@ -1,25 +1,25 @@
 const windshaft = require('windshaft');
 const MapConfig = windshaft.model.MapConfig;
 const Datasource = windshaft.model.Datasource;
-const cleanUpQueryParams = require('../../middleware/clean-up-query-params');
-const credentials = require('../../middleware/credentials');
-const dbConnSetup = require('../../middleware/db-conn-setup');
-const authorize = require('../../middleware/authorize');
+const cleanUpQueryParams = require('../middlewares/clean-up-query-params');
+const credentials = require('../middlewares/credentials');
+const dbConnSetup = require('../middlewares/db-conn-setup');
+const authorize = require('../middlewares/authorize');
 const initProfiler = require('./middlewares/init-profiler');
 const checkJsonContentType = require('./middlewares/check-json-content-type');
 const incrementMapViewCount = require('./middlewares/increment-map-view-count');
 const augmentLayergroupData = require('./middlewares/augment-layergroup-data');
-const cacheControlHeader = require('../../middleware/cache-control-header');
-const cacheChannelHeader = require('../../middleware/cache-channel-header');
-const surrogateKeyHeader = require('../../middleware/surrogate-key-header');
-const lastModifiedHeader = require('../../middleware/last-modified-header');
+const cacheControlHeader = require('../middlewares/cache-control-header');
+const cacheChannelHeader = require('../middlewares/cache-channel-header');
+const surrogateKeyHeader = require('../middlewares/surrogate-key-header');
+const lastModifiedHeader = require('../middlewares/last-modified-header');
 const lastUpdatedTimeLayergroup = require('./middlewares/last-updated-time-layergroup');
 const layerStats = require('./middlewares/layer-stats');
 const layergroupIdHeader = require('./middlewares/layergroup-id-header');
 const layergroupMetadata = require('./middlewares/layergroup-metadata');
 const mapError = require('./middlewares/map-error');
 const CreateLayergroupMapConfigProvider = require('../../models/mapconfig/provider/create-layergroup-provider');
-const rateLimit = require('../../middleware/rate-limit');
+const rateLimit = require('../middlewares/rate-limit');
 const { RATE_LIMIT_ENDPOINTS_GROUPS } = rateLimit;
 
 /**
