+++ lib/cartodb/routers/middlewares/clean-up-query-params.js
