+++ lib/cartodb/routers/middlewares/cors.js
