+++ lib/cartodb/routers/middlewares/credentials.js
