+++ lib/cartodb/routers/middlewares/logger.js
