+++ lib/cartodb/routers/middlewares/send-response.js
