+++ lib/cartodb/routers/middlewares/served-by-host-header.js
