+++ lib/cartodb/routers/middlewares/stats.js
@@ -1,4 +1,4 @@
-const Profiler = require('../stats/profiler_proxy');
+const Profiler = require('../../stats/profiler_proxy');
 const debug = require('debug')('windshaft:cartodb:stats');
 const onHeaders = require('on-headers');
