+++ lib/cartodb/routers/middlewares/surrogate-key-header.js
@@ -1,5 +1,5 @@
-const NamedMapsCacheEntry = require('../cache/model/named_maps_entry');
-const NamedMapMapConfigProvider = require('../models/mapconfig/provider/named-map-provider');
+const NamedMapsCacheEntry = require('../../cache/model/named_maps_entry');
+const NamedMapMapConfigProvider = require('../../models/mapconfig/provider/named-map-provider');
 
 module.exports = function setSurrogateKeyHeader ({ surrogateKeysCache }) {
     return function setSurrogateKeyHeaderMiddleware(req, res, next) {
