+++ lib/cartodb/routers/middlewares/syntax-error.js
