+++ lib/cartodb/routers/middlewares/user.js
@@ -1,4 +1,4 @@
-const CdbRequest = require('../models/cdb_request');
+const CdbRequest = require('../../models/cdb_request');
 
 module.exports = function user () {
     const cdbRequest = new CdbRequest();
