+++ lib/cartodb/routers/middlewares/vector-error.js
@@ -1,5 +1,5 @@
 const fs = require('fs');
-const timeoutErrorVectorTile = fs.readFileSync(__dirname + '/../../../assets/render-timeout-fallback.mvt');
+const timeoutErrorVectorTile = fs.readFileSync(__dirname + '/../../../../assets/render-timeout-fallback.mvt');
 
 module.exports = function vectorError() {
     return function vectorErrorMiddleware(err, req, res, next) {
