+++ lib/cartodb/routers/template/admin.js
@@ -1,6 +1,6 @@
 const { templateName } = require('../../backends/template_maps');
-const credentials = require('../../middleware/credentials');
-const rateLimit = require('../../middleware/rate-limit');
+const credentials = require('../middlewares/credentials');
+const rateLimit = require('../middlewares/rate-limit');
 const { RATE_LIMIT_ENDPOINTS_GROUPS } = rateLimit;
 
 /**
