+++ lib/cartodb/routers/template/tile.js
@@ -1,14 +1,14 @@
-const cleanUpQueryParams = require('../../middleware/clean-up-query-params');
-const credentials = require('../../middleware/credentials');
-const dbConnSetup = require('../../middleware/db-conn-setup');
-const authorize = require('../../middleware/authorize');
+const cleanUpQueryParams = require('../middlewares/clean-up-query-params');
+const credentials = require('../middlewares/credentials');
+const dbConnSetup = require('../middlewares/db-conn-setup');
+const authorize = require('../middlewares/authorize');
 const namedMapProvider = require('./middlewares/named-map-provider');
-const cacheControlHeader = require('../../middleware/cache-control-header');
-const cacheChannelHeader = require('../../middleware/cache-channel-header');
-const surrogateKeyHeader = require('../../middleware/surrogate-key-header');
-const lastModifiedHeader = require('../../middleware/last-modified-header');
-const vectorError = require('../../middleware/vector-error');
-const rateLimit = require('../../middleware/rate-limit');
+const cacheControlHeader = require('../middlewares/cache-control-header');
+const cacheChannelHeader = require('../middlewares/cache-channel-header');
+const surrogateKeyHeader = require('../middlewares/surrogate-key-header');
+const lastModifiedHeader = require('../middlewares/last-modified-header');
+const vectorError = require('../middlewares/vector-error');
+const rateLimit = require('../middlewares/rate-limit');
 const { RATE_LIMIT_ENDPOINTS_GROUPS } = rateLimit;
 
 function TileTemplateController (
